Fixed possible division by zero error
If for some reason the timestamp delta is zero, presumably because the
check has been called twice at the exact same time, various rate
calculations caused a division by zero error.

We now check for zero in delta_timestamp and assume a rate of 0 for the
corresponding rate as a zero delta_timestamp implies no time has passed
so the value for which the rate per time is being calculated could not
have changed either.